Assessment: submit uploaded submissions and split archive parsing
Address the batch upload review feedback and the failing Server Code
Quality and Client Tests checks:

- Mark the external submission created for a participant without a prior
  submission as submitted and dated, so the imported assessment is no
  longer omitted from submitted-result queries and views (mirrors the
  external-submission path in ParticipationService). Add a regression test.
- Extract the ZIP/CSV parsing into AssessmentUploadArchiveParser so
  AssessmentUploadService drops back under the 1000-line large-class limit.
- Use the tum-ui-dialog semantic `size` input and `shown` output in the
  upload dialog instead of host `[style]` and the removed `(onShow)`.
